+++ protected/controllers/SpecialityController.php
@@ -4,6 +4,30 @@ class SpecialityController extends Controller
 {
     public $name = "Specialities";
 
+/*
+    public function filters()
+    {
+        return array(
+            'accessControl',
+        );
+    }
+
+    public function accessRules()
+    {
+        return array(
+            array('allow',
+                'actions'=>array('create', 'delete', 'update'),
+                'roles'=>array('dephead'),
+            ),
+
+            //array('deny',
+            //    'actions'=>array('update', 'create', 'delete'),
+            //    'users'=>array('*'),
+            //),
+        );
+    }
+*/
+
     /**
      * Displays a particular model.
      * @param integer $id the ID of the model to be displayed
@@ -21,6 +45,10 @@ class SpecialityController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::app()->user->checkAccess('manageSpeciality'))
+        {
+            throw new CHttpException(403, Yii::t('yii','You are not authorized to perform this action.'));
+        }
         $model = new Speciality;
 
         if (isset($_POST['Speciality'])) {
@@ -43,6 +71,16 @@ class SpecialityController extends Controller
     {
         $model = Speciality::model()->loadContent($id);
 
+        if(!Yii::app()->user->checkAccess('manageSpeciality',
+            array(
+                'id' => $model->department->head_id,
+                'type' => User::TYPE_TEACHER,
+            )
+        ))
+        {
+            throw new CHttpException(403, Yii::t('yii','You are not authorized to perform this action.'));
+        }
+
         if (isset($_POST['Speciality'])) {
             $model->attributes = $_POST['Speciality'];
             if ($model->save())
@@ -61,6 +99,10 @@ class SpecialityController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::app()->user->checkAccess('manageSpeciality'))
+        {
+            throw new CHttpException(403, Yii::t('yii','You are not authorized to perform this action.'));
+        }
         Speciality::model()->loadContent($id)->delete();
 
         if (!isset($_GET['ajax']))
